+++ Tests/Help.Tests.ps1
@@ -1,19 +1,15 @@
 ﻿# https://lazywinadmin.com/2016/05/using-pester-to-test-your-comment-based.html
 
-$ProjectRoot = Resolve-Path "$($PSScriptRoot)/.."
-$ModuleRoot = Split-Path (Resolve-Path "$($ProjectRoot)/*/*.psm1")
-$ModuleName = Split-Path $ModuleRoot -Leaf
-
-Import-Module $ModuleRoot
+Import-Module "$($env:BHModulePath)\$($env:BHProjectName).psm1"
 
 Describe "$($ModuleName) Comment Based Help" -Tags "Module" {
-    $FunctionsList = (Get-ChildItem "$($ModuleRoot)\Public").BaseName
+    $FunctionsList = (Get-ChildItem -Path "$($env:BHModulePath)\Public").BaseName
 
     foreach ($Function in $FunctionsList) {
         # Retrieve the Help of the function
         $Help = Get-Help -Name $Function -Full
 
-        $Notes = ($Help.alertSet.Alert.Text -split '\n')
+        $Notes = ($Help.AlertSet.Alert.Text -split '\n')
 
         # Parse the function using AST
         $AST = [System.Management.Automation.Language.Parser]::ParseInput((Get-Content function:$Function), [ref]$null, [ref]$null)
@@ -33,14 +29,14 @@ Describe "$($ModuleName) Comment Based Help" -Tags "Module" {
             $ASTParameters = $ast.ParamBlock.Parameters.Name.VariablePath.UserPath
 
             It "Parameter - Compare Count Help/AST" {
-                $HelpParameters.Name.Count -eq $ASTParameters.Count | Should Be $true
+                $HelpParameters.Name.Count -eq $ASTParameters.Count | Should -Be $true
             }
 
             # Parameter Description
             If (-not [String]::IsNullOrEmpty($ASTParameters)) {
                 # IF ASTParameters are found
                 $HelpParameters | ForEach-Object {
-                    It "Parameter $($_.Name) - Should contains description" {
+                    It "Parameter $($_.Name) - Should contain description" {
                         $_.Description | Should -not -BeNullOrEmpty
                     }
                 }
@@ -48,7 +44,7 @@ Describe "$($ModuleName) Comment Based Help" -Tags "Module" {
 
             # Examples
             it "Example - Count should be greater than 0" {
-                $Help.Examples.Example.Code.Count | Should BeGreaterthan 0
+                $Help.Examples.Example.Code.Count | Should -BeGreaterThan 0
             }
 
             # Examples - Remarks (small description that comes with the example)
@@ -57,6 +53,6 @@ Describe "$($ModuleName) Comment Based Help" -Tags "Module" {
                     $Example.Remarks | Should -not -BeNullOrEmpty
                 }
             }
-        }
+        } # Context: $Function - Help
     } # foreach
 } # Describe